add noembed buildtag
works like so:
```
$ make install BUILDTAGS="seccomp noembed"
+ install
go install -a -tags "seccomp noembed" -ldflags "-w -X github.com/genuinetools/img/version.GITCOMMIT=455629ae-dirty -X github.com/genuinetools/img/version.VERSION=v0.4.1" .

$ img build -t thing .
FATA[0000] Installing embedded runc binary failed: please install `runc`
```

closes #107
